# src/main/resources/openapi-draft.yaml
openapi: 3.0.1
info:
  title: Customer API
  version: 1.0.0
  description: Customer management API

paths:
  /api/v1/customers:
    get:
      summary: Get all customers
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
    post:
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        201:
          description: Created
        400:
          description: Validation error
        409:
          description: Duplicate email

  /api/v1/customers/{id}:
    get:
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success
        404:
          description: Not found
    put:
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        200:
          description: Updated
        404:
          description: Not found
    delete:
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Deleted
        404:
          description: Not found

components:
  schemas:
    CustomerRequest:
      type: object
      required:
        - name
        - email
        - phone
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@email.com"
        phone:
          type: string
          example: "1234567890"
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "john@email.com"
        phone:
          type: string
          example: "1234567890"